Error Fixes
I fixed the errors just now, I'll be pushing the fixed changes to the GitHub repository.

Error 1: Duplicates
AddCar request now does a pre-check where it searched for the car before adding it to the database, if Car already exists it'll return the car's data.

Error 2: Errors field with no error
The reason an error is not shown is because it's not technically an error it was due to cars not existing which indicated the request went through however just received nothing.
When this happens I manually set the success variable to false. However, this is not meant to be the case and I have changed that.

Error 3: SELECT * in SQL Requests
I have fixed this and instead of SELECT *, I have replaced that with the actual table columns for both the Cars and Dealers Tables for performance and scalability.